Add preview flags (#37)
* Remove redundant error print

* Add flag support for preview command

* Add dry-run support

+++ internal/file/replace.go
@@ -5,11 +5,36 @@ import (
 	"strings"
 )
 
+type replaceMode struct {
+	isDryRun bool
+	isForce  bool
+}
+type ReplaceOption func(*replaceMode) error
+
+func WithDryRun() ReplaceOption {
+	return func(m *replaceMode) error {
+		m.isDryRun = true
+		return nil
+	}
+}
+
+func WithForce() ReplaceOption {
+	return func(m *replaceMode) error {
+		m.isForce = true
+		return nil
+	}
+}
+
 // ReplaceWithAfter replaces the original file content with the processed
 // content. This is done by creating a temp file first, and replacing it.
 //
 // TODO: Ensure file mode is kept, or clarify in the comment.
-func (f *File) ReplaceWithAfter() error {
+func (f *File) ReplaceWithAfter(options ...ReplaceOption) error {
+	mode := &replaceMode{}
+	for _, opt := range options {
+		opt(mode)
+	}
+
 	file, err := os.CreateTemp("/tmp/", "importer_replace_*")
 	if err != nil {
 		return err // TODO: test coverage
@@ -19,6 +44,11 @@ func (f *File) ReplaceWithAfter() error {
 	_, err = file.Write(f.ContentAfter)
 	if err != nil {
 		return err // TODO: test coverage
+	}
+
+	// Exit early if dry-run
+	if mode.isDryRun {
+		return nil
 	}
 
 	err = os.Rename(file.Name(), f.FileName)
@@ -33,7 +63,12 @@ func (f *File) ReplaceWithAfter() error {
 // content. This is done by creating a temp file first, and replacing it.
 //
 // TODO: Ensure file mode is kept, or clarify in the comment.
-func (f *File) ReplaceWithPurged() error {
+func (f *File) ReplaceWithPurged(options ...ReplaceOption) error {
+	mode := &replaceMode{}
+	for _, opt := range options {
+		opt(mode)
+	}
+
 	file, err := os.CreateTemp("/tmp/", "importer_replace_*")
 	if err != nil {
 		return err // TODO: test coverage
@@ -45,6 +80,11 @@ func (f *File) ReplaceWithPurged() error {
 	_, err = file.WriteString(data)
 	if err != nil {
 		return err // TODO: test coverage
+	}
+
+	// Exit early if dry-run
+	if mode.isDryRun {
+		return nil
 	}
 
 	err = os.Rename(file.Name(), f.FileName)
